Put listing query tests in a single file
